refactor(assets): extract asset-media + asset-trash from asset-service
WS6 god-file split, following the audit's seam map. Two self-contained,
cycle-free modules move out of the 835-line asset-service.ts:

- lib/asset-media.ts — the lazy `sharp` loader + imageDimensions +
  generateThumbnail (ffmpeg fallback). This is the only module-level mutable
  state in the asset service (the sharpModule promise cache), so isolating it
  keeps the create/read/mutation paths side-effect-free. The 3 helpers were
  module-private, so no public surface changes.
- lib/asset-trash.ts — soft-delete (deleteAsset) + list/restore/permanently-
  delete/empty. Self-contained: depends only on asset-id, asset-lock, manifest,
  and the task-asset index. Trash ops deliberately bypass the manifest-write
  choke point (a rename, not a write), so they evict/relink the task-asset index
  explicitly — those calls travel with the functions.

`assetDirAbs` moves to its natural home, asset-id.ts (it's pure assetId → dir
resolution). Both asset-service (mutations) and asset-trash import it there — no
cycle, since asset-service never imports asset-trash.

The two trash consumers (plugins/assets/index.ts, routes/versioned.ts) and 4
test files are repointed to asset-trash. The two fragile non-TS-caught importers
the audit flagged — plugin-context-services.ts (string-literal dynamic import)
and post-channel.ts (path-alias) — use only create/read functions and are
untouched.

asset-service.ts: 835 -> 682 lines. Pure relocation.

Deferred: asset-mutations + asset-upsert split and the optional redesigns
(mutateManifest combinator, iterateStoreManifests walker, deduping the
images-plugin sharp copy against the new asset-media).

Verified: typecheck + lint clean; assets suite 242/0 (lifecycle/manifest-cache/
tag-ops/list-by-task all exercise the moved trash + media code); full suite
5123/0; binary build; isolated-home boot smoke (Assets activates; GET
/api/plugins/assets/trash + /versioned both 200).
